fix: use github action secret
